test(sdk-core): Add assets E2E tests 🧪
