+++ packages/sdk-core/e2e/index.ts
@@ -1,0 +1,3 @@
+export * from './utils'
+export * from './generateE2eTests'
+export * from './generateAssetsTests'
